+++ src/main/java/calculator/Expression.java
@@ -1,11 +1,11 @@
 package calculator;
 
 public class Expression {
-    int num1;
-    int num2;
+    double num1;
+    double num2;
     char operator;
 
-    public Expression (int num1, int num2, char operator) {
+    public Expression (double num1, double num2, char operator) {
         this.num1 = num1;
         this.num2 = num2;
         this.operator = operator;
